splitted huge file into severals
