+++ build.sbt
@@ -16,5 +16,7 @@ libraryDependencies ++= Seq(
   "com.assembla.scala-incubator" % "graph-core_2.10" % "1.7.3",
   "org.apache.jena" % "jena-arq" % "2.11.1" ,
   "org.scalatest" % "scalatest_2.10" % "2.1.0-RC2",
-  "junit" % "junit" % "4.10" % "test")
+  "junit" % "junit" % "4.10" % "test",
+  "org.openrdf.sesame" % "sesame-model" % "2.7.10"
+)
