+++ src/main/scala/es/weso/jena/JenaMapper.scala
@@ -48,17 +48,16 @@ trait JenaMapper {
        	m.createResource(i.str)
      case StringLiteral(str) 				 => 
        	m.createLiteral(str,false)
-     case DatatypeLiteral(str,i:IRI) => {
-        i.str match {
-          case `xsdinteger` => m.createTypedLiteral(str,XSDDatatype.XSDinteger) 
-          case `xsddouble` => m.createTypedLiteral(str,XSDDatatype.XSDdouble)
-          case `xsddecimal` => m.createTypedLiteral(str,XSDDatatype.XSDdecimal)
-          case `xsdboolean` => m.createTypedLiteral(str,XSDDatatype.XSDboolean)
-          case _ => m.createTypedLiteral(str,new BaseDatatype(i.str))
-        }
-     }
+     case DatatypeLiteral(str,i:IRI) =>
+       i.str match {
+         case `xsdinteger` => m.createTypedLiteral(str,XSDDatatype.XSDinteger)
+         case `xsddouble` => m.createTypedLiteral(str,XSDDatatype.XSDdouble)
+         case `xsddecimal` => m.createTypedLiteral(str,XSDDatatype.XSDdecimal)
+         case `xsdboolean` => m.createTypedLiteral(str,XSDDatatype.XSDboolean)
+         case _ => m.createTypedLiteral(str,new BaseDatatype(i.str))
+       }
      case DecimalLiteral(d) 		=> 
-     	m.createTypedLiteral(d.toString,XSDDatatype.XSDdecimal)
+     	m.createTypedLiteral(d.toString(),XSDDatatype.XSDdecimal)
      case IntegerLiteral(i) 		=> 
        	m.createTypedLiteral(i.toString,XSDDatatype.XSDinteger)
      case LangLiteral(l,Lang(lang)) => m.createLiteral(l,lang)
