+++ src/main/scala/es/weso/rdfgraph/RDFGraph.scala
@@ -3,9 +3,8 @@ package es.weso.rdfgraph
 import es.weso.tgraph._
 import es.weso.tgraph.TGraph
 import es.weso.tgraph.TGraphImpl
-
 import scala.collection.Set
-import scala.collection.immutable.Map
+import scala.collection.immutable.{Map}
 import scalax.collection.immutable.Graph
 import es.weso.rdfgraph.statements.RDFTriple
 import es.weso.rdfgraph.nodes._
@@ -104,173 +103,6 @@ abstract class RDFGraph {
   
 }
 
-case class Ground(graph : TGraph[RDFNode])
-                 (implicit seed : BNodeId) extends RDFGraph {
-
-  override def isEmpty = graph.isEmpty
-  
-
-  override def insertTriple(triple: RDFTriple) : RDFGraph = {
-     Ground(graph.addTriple(triple.subj,triple.pred,triple.obj))
-  }
-  
-  def getBNodeMap(node : RDFNode, map: Map[BNodeId,BNodeId]) : RDFNode = {
-    node match {
-      case b@BNodeId(_) => map(b)
-      case other => other
-    }
-  }
-
-  override def insertTripleMap(triple: RDFTriple,
-		  			  map : Map[BNodeId,BNodeId]) : RDFGraph = {
-     val s = getBNodeMap(triple.subj,map)
-     val p = triple.pred
-     val o = getBNodeMap(triple.obj,map)
-     Ground(graph.addTriple(s,p,o))
-  }
-
-  
-  override def addTriplesBNodes(
-		  bnodes : Set[BNodeId], 
-		  triples : Set[RDFTriple],
-		  map : Map[BNodeId,BNodeId]) : RDFGraph = {
-    if (bnodes.isEmpty){
-      val current : RDFGraph = this
-      triples.foldLeft (current) ((g,triple) => 
-        						g.insertTripleMap(triple,map))
-    } else {
-      Exists((bnode) => addTriplesBNodes(bnodes.tail,
-    		  					triples,
-    		  					map + (bnodes.head -> bnode) )) 
-    }
-  }
-
-  /**
-   * addTriples inserts a set of triples into a graph 
-   * 
-   * It takes control of possible bnodes in the triples renaming them 
-   * @param triples set of triples
-   * 
-   */
-  override def addTriples(triples : Set[RDFTriple]) : RDFGraph = {
-    val bnodes = RDFTriple.collectBNodes(triples)
-    addTriplesBNodes(bnodes,triples,Map.empty)
-  }
-  
-  override def IRIs : Set[IRI] = {
-    graph.nodes.filter(_.isIRI).map(_.toIRI)
-  }
-  
-  /** get the triples of a graph
-   * 
-   * @param seed represents the seed for blank node identifier generation
-   * (default value = 0)
-   * 
-   * Ground graphs ignore the seed parameter
-   * 
-   */
-  override def triples(implicit seed : BNodeId) : Set[RDFTriple] = {
-    graph.triples.map((t) => RDFTriple(t._1,t._2.toIRI,t._3))
-  }  
-  
-  override def merge (other: RDFGraph) : RDFGraph = {
-    val g = this.addTriples(other.triples)
-    g
-  }
-  
-  override def show (implicit seed : BNodeId): String = {
-    this.toString
-  }
-
-  /**
-   * Decompose a graph from a given IRI
-   * @param node resource from which we are decomposing the graph
-   * 
-   * @author labra
-   */
-  def decomp (node : IRI) : Option[(TContext[RDFNode],RDFGraph)] = {
-    graph.decomp(node) match {
-      case None => None
-      case Some((ctx,g)) => Some((ctx,Ground(g)))
-    }
-  }
-  
-  override def foldRDFGraphSeed[A] (e: A, f:(TContext[RDFNode],A) => A, seed : BNodeId) : A = {
-    graph.foldTGraph(e)(f)
-  }
-
-  def foldRDFGraphSeedOrd[A] (
-		  e: A, 
-		  f: (TContext[RDFNode],A) => A,
-		  seed : BNodeId)
-		  (implicit ord : Ordering[RDFNode]) : A = {
-    graph.foldTGraphOrd(e)(f)(ord)
-  }
-    
-}
-
-case class Exists(fn : BNodeId => RDFGraph)
-				   (implicit seed : BNodeId) extends RDFGraph {
-
-  override def isEmpty = false
-  
-  override def insertTriple(triple: RDFTriple) : RDFGraph = {
-    Exists((bnode) => (fn(bnode)).insertTriple(triple))
-  }
-  
-  override def addTriples(triples : Set[RDFTriple]) : RDFGraph = {
-    Exists((bnode) => (fn(bnode)).addTriples(triples))
-  }
-
-  override def triples (implicit seed : BNodeId): Set[RDFTriple] = {
-   (fn(seed)).triples(seed.newBNodeId)  
-  }
-
-  override def IRIs: Set[IRI] = {
-   (fn(seed)).IRIs  
-  }
-
-  /*
-   * merges this graph with another one
-   */
-  override def merge (other : RDFGraph) : RDFGraph = {
-    Exists((bnode => (fn(bnode)).merge(other)))
-  }
-
-  /*
-   * add triples which can have a set of bNodes
-   */
-  override def addTriplesBNodes(bnodes : Set[BNodeId], 
-		  triples : Set[RDFTriple],
-		  map : Map[BNodeId,BNodeId]) : RDFGraph = {
-    Exists((bnode) => (fn(bnode)).addTriplesBNodes(bnodes,triples,map))
-  }
-  
-  override def insertTripleMap(triple: RDFTriple,
-		  			  map : Map[BNodeId,BNodeId]) : RDFGraph = {
-    Exists((bnode) => (fn(bnode)).insertTripleMap(triple,map))
-  }
-  
-  override def show(implicit seed : BNodeId) : String = {
-    "Exists " + seed.id + " ( " + (fn(seed)).show(seed.newBNodeId) + ")"
-  }
-  
-  override def foldRDFGraphSeed[A] (e: A, f: (TContext[RDFNode],A) => A, seed: BNodeId) : A = {
-    (fn(seed)).foldRDFGraphSeed(e,f,seed.newBNodeId)
-  }
-
-  def foldRDFGraphSeedOrd[A] (
-		  e: A, 
-		  f: (TContext[RDFNode],A) => A,
-		  seed : BNodeId)
-		  (implicit ord : Ordering[RDFNode]) : A = {
-    (fn(seed)).foldRDFGraphSeedOrd(e,f,seed.newBNodeId)(ord)
-  }
-  
-}
-
-
-
 object RDFGraph {
 
   // Implicit definition for starting BNodeId
@@ -294,7 +126,7 @@ object RDFGraph {
   def showFolds (g : RDFGraph) : String = {
     g.foldRDFGraph(
          "\n", 
-         ((ctx : TContext[RDFNode],r : String) => "ctx: " + ctx + "\n" + r )
+         (ctx: TContext[RDFNode], r: String) => "ctx: " + ctx + "\n" + r
          )
   }
 
@@ -318,7 +150,7 @@ object RDFGraph {
           case _ => throw new RDFNodeException("Unexpected values " + (x,y) + " comparing RDFNodes")
        }
    	}
-   	g.foldRDFGraphOrd("", ((ctx : TContext[RDFNode],r : String) => "ctx: " + ctx + "\n" + r ))
+   	g.foldRDFGraphOrd("", (ctx: TContext[RDFNode], r: String) => "ctx: " + ctx + "\n" + r)
   }
