+++ src/main/scala/es/weso/sesame/SesameGraph.scala
@@ -1,0 +1,10 @@
+package es.weso.sesame
+
+import es.weso.rdfgraph.RDFGraph
+
+/**
+ * Created by antonkulaga on 2/25/14.
+ */
+//class SesameGraph extends RDFGraph{
+//
+//}
