+++ src/test/scala/es/weso/qgraph/QGraphSuite.scala
@@ -3,10 +3,6 @@ package es.weso.qgraph
 import org.scalatest.FunSuite
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
-import es.weso.tgraph.TGraph._
-import es.weso.tgraph.TGraphImpl._
-import scalax.collection.immutable.Graph
-import scalax.collection.GraphEdge._
 import org.scalatest.Matchers
 
 @RunWith(classOf[JUnitRunner])
